+++ src/io/wav_file.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 #include <vector>
 #include <cstdint> // for some data types like uint32_t
-#include<fstream>
+#include <fstream>
 
 /* 
     used for writing bytes in the Write Function 
@@ -11,7 +11,7 @@
 */
 
 // TODO: make the function checking for Endianness and handle writing on big or little Endianness
-bool write_as_bytes(std::ofstream &file, const void *val, const int byte_size){
+bool write_as_bytes(std::ofstream &file, const void *val, const size_t byte_size){
     return static_cast<bool>(
         file.write(
             (reinterpret_cast<const char*>(val)), byte_size
